Rename package to crypto
