+++ modules/ballerina-test/src/test/java/org/ballerinalang/test/nativeimpl/functions/UtilTest.java
@@ -35,7 +35,7 @@ public class UtilTest {
 
     @BeforeClass
     public void setup() {
-        compileResult = BCompileUtil.compile(this, "test-src", "nativeimpl/functions/utils-test.bal");
+        compileResult = BCompileUtil.compile(this, "test-src", "nativeimpl/functions/util-test.bal");
     }
 
     @Test
