+++ modules/ballerina-test/src/test/resources/test-src/nativeimpl/functions/crypto-test.bal
@@ -1,9 +1,9 @@
-import ballerina.security;
+import ballerina.crypto;
 
 function testHmac (string base, string key, string algo) (string) {
-    return security:getHmac(base, key, algo);
+    return crypto:getHmac(base, key, algo);
 }
 
 function testMessageDigest (string base, string algo) (string) {
-    return security:getHash(base, algo);
+    return crypto:getHash(base, algo);
 }
