+++ modules/ballerina-test/src/test/resources/test-src/nativeimpl/functions/util-test.bal
@@ -1,5 +1,5 @@
 import ballerina.util;
-import ballerina.security;
+import ballerina.crypto;
 
 function testEncodeDecode (string s) (string) {
     return util:base64Decode(util:base64Encode(s));
@@ -34,9 +34,9 @@ function testBase64ToBase16Encoding (string s) (string) {
 }
 
 function testHMACValueFromBase16ToBase64Encoding (string base, string key, string algo) (string) {
-    return util:base16ToBase64Encode(security:getHmac(base, key, algo));
+    return util:base16ToBase64Encode(crypto:getHmac(base, key, algo));
 }
 
 function testHMACValueFromBase64ToBase16Encoding (string base, string key, string algo) (string) {
-    return util:base64ToBase16Encode(util:base16ToBase64Encode(security:getHmac(base, key, algo)));
+    return util:base64ToBase16Encode(util:base16ToBase64Encode(crypto:getHmac(base, key, algo)));
 }
